New Medical Page Format
Everything is now JSON based. Starts with building a JSON file for each individual with personal information. Then include an entry in Master.json for that person.

Labs.pl (in ../../Labs) is used to create a JSON of blood test data. Outputs to [DKF|JMF]_Labs.json.

Finally, medical excel spreadsheet exports medicine lists for each individual. This is a TSV file.